requirements.txt
python-telegram-bot==21.5
feedparser==6.0.11
openai==1.51.0
requests==2.32.3
python-dotenv==1.0.1
